+++ demos/53-circle-sector-intersections/index.js
@@ -17,7 +17,6 @@
 
   // Fetch the GET params
   let GUP = gup();
-  var isDarkmode = detectDarkMode(GUP);
   window.addEventListener("load", function () {
     var isDarkmode = detectDarkMode(GUP);
     var params = new Params(GUP);