+++ demos/extend-lilgui-colorwithalpha.js
@@ -116,6 +116,7 @@ lil.GUI.prototype.addColorWithAlpha = function (object, colorProperty, alphaProp
   cntrlr.title = cntrlr.baseColorController.title;
   cntrlr.enable = cntrlr.baseColorController.enable;
   cntrlr.disable = cntrlr.baseColorController.disable;
+  cntrlr.listen = cntrlr.baseColorController.listen;
 
   return cntrlr;
 };