+++ src/esm/utils/gup.d.ts
@@ -9,5 +9,6 @@
  * Ported to typescript.
  * @date 2021-05-21
  * @modified 2024-03-10 Fixed some type for Typescript 5 compatibility.
+ * @modified 2024-08-26 Decoding URI components in GET params.
  */
 export declare const gup: () => Record<string, string>;