+++ src/esm/utils/gup.js
@@ -9,6 +9,7 @@
  * Ported to typescript.
  * @date 2021-05-21
  * @modified 2024-03-10 Fixed some type for Typescript 5 compatibility.
+ * @modified 2024-08-26 Decoding URI components in GET params.
  */
 // Get the URI GET params as an assoc.
 //
@@ -18,7 +19,8 @@
 export const gup = () => {
     const vars = {};
     globalThis.location.href.replace(/[?&]+([^=&]+)=([^&]*)/gi, (_m, key, value) => {
-        return (vars[key] = value);
+        const keyName = key;
+        return (vars[decodeURIComponent(keyName)] = value);
     });
     return vars;
 };